0x0F-python-object_relational_mapping

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,36 +1,19 @@
 #!/usr/bin/python3
-"""
-Script that deletes all State objects with name containing letter 'a'
-from the database hbtn_0e_6_usa.
+"""Script that deletes all State objects with a name containing the letter a"""
 
-Usage: ./13-model_state_delete_a.py <mysql username> \
-                                   <mysql password> \
-                                   <database name>
-"""
-
-from model_state import Base, State
+import sys
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from sys import argv
-
+from model_state import State
 
 if __name__ == "__main__":
-    # Create connection to database
-    engine = create_engine('mysql+mysqldb://{}:{}@localhost/{}'.format(
-        argv[1], argv[2], argv[3]), pool_pre_ping=True)
-
-    # Create session factory
+    engine = create_engine("mysql+mysqldb://{}:{}@localhost/{}"
+                           .format(sys.argv[1], sys.argv[2], sys.argv[3]),
+                           pool_pre_ping=True)
     Session = sessionmaker(bind=engine)
-
-    # Create session
     session = Session()
-    Base.metadata.create_all(engine)
 
-    # Delete states containing letter 'a'
-    states = session.query(State).filter(State.name.like('%a%')).all()
-    for state in states:
-        session.delete(state)
-
-    # Commit changes and close session
+    for state in session.query(State):
+        if "a" in state.name:
+            session.delete(state)
     session.commit()
-    session.close()
